refactored status compnent of ticket section
